Establish basis for modularizing VM construction
This allows backends to further specialize VM construction to meet their
own needs without having to duplicate any VM construction code in any of
the VM-based backends.

A backend can participate in the construction of a VM by implementing
the vm_base::builder::Hook trait and overriding some (or all) of the
methods in that trait to perform the work it needs to do.

The vm_base::builder::Hook trait can be easily extended to add additional
steps if needed.

No functional changes in terms of the VM's personality; this is just a
large refactoring.

Unused abstractions, such as the Allocator type from cpu.rs, have been
removed. Furthermore, some things that were needlessly coupled, such as
the VMSetup struct were also removed (since this type is only used
during 1-time setup).