shim-sev: handle multiple return values for syscalls
`rdx` is sometimes used as a secondary return value.

Because `rdx` is not marked as `clobber` for the musl syscall macro,
the original value has to be returned, if `rdx` is unused as a return
value.

To combat the use of the stack in `_syscall_enter` with `opt-level=0`,
reserve a little red zone on the stack for `#[cfg(debug_assertions)]`.
In `opt-level=3` the compiler does emit the correct assembly code
without using the stack.

This patch also adds a dummy path for the enarx syscalls, which can make
use of the second return value.

Signed-off-by: Harald Hoyer <harald@redhat.com>