shim-sev: handle multiple return values for syscalls
`rdx` is sometimes used as a secondary return value.

This patch lets `syscall_rust` return a double u64 struct instead of u64,
which will use `rax` for the first 64 bits and `rdx` for the last 64 bit.

Because `rdx` is not marked as `clobber` for the musl syscall macro,
the original value has to be returned, if `rdx` is unused as a return
value.

This patch also adds a dummy path for the enarx syscalls, which can make
use of the second return value.

Signed-off-by: Harald Hoyer <harald@redhat.com>