fix(sallyport): Fix the sallyport block size
The sallyport block size was set 512 * 4096 bytes, whereas
the space was never allocated in the kvm backend.

This patch sets the capacity to hold at least a maximum UDP packet.

When multi-threading is implemented multiple buffers need to be
available for concurrent syscalls.
Related: https://github.com/enarx/enarx-keepldr/issues/156

Fixes: https://github.com/enarx/enarx-keepldr/issues/23

Signed-off-by: Harald Hoyer <harald@redhat.com>