feat(shim-sev): Use multiple sallyport::Block
Because the setup area is limited in memory size, not each thread can
get its own `sallyport::Block` to communicate with the host.

Therefore a pool of blocks is reserved, which can be dynamically used
by the shim indicating the actual block via the value passed in the
IOPort.

A future extension could reserve additional blocks via a enarx specific
syscall.

This patch also removes the now obsolete `backend::kvm::vm::Builder::Config`.

Fixes: https://github.com/enarx/enarx-keepldr/issues/156

Signed-off-by: Harald Hoyer <harald@redhat.com>