fix: update dependencies for cargo deny
cargo deny complains about smallvec 1.5.1 having a security
vulnerability.

ID: RUSTSEC-2021-0003
Advisory: https://rustsec.org/advisories/RUSTSEC-2021-0003
A bug in the `SmallVec::insert_many` method caused it to allocate a buffer
that was smaller than needed.  It then wrote past the end of the buffer,
causing a buffer overflow and memory corruption on the heap.

A manual `cargo update` updates the deps of the deps, so a fixed version
is used in the end.

Signed-off-by: Harald Hoyer <harald@redhat.com>