Rework sallyport exit handling
Before this commit we had a custom system where we would suspend
execution of the thread in CSSA 1 to perform any action over sallyport.
This meant we had tons of assembly code to make this all work.

This commit simplifies how this all works. The new system is described
below.

First, normal code executes in CSSA 0. When it hits an instruction SGX
can't handle (i.e. syscall or cpuid), we have an AEX which suspends
execution and exits to the host. The host then reenters the enclave in
CSSA 1.

Second, in CSSA 1 we evaluate the exit reason from CSSA 0. This is
usually a `cpuid` or `syscall` instruction. We then prepare the
sallyport for execution. Instead of triggering our custom assembly,
now we execute a `syscall` instruction again. This causes an AEX which
suspends execution from CSSA 1 and exits to the host. The host then
reenters the enclave in CSSA 2.

Third, in CSSA 2 we simply skip the `syscall` instruction in CSSA 1 and
exit successfully. Upon a successful return from CSSA 2, the host
executes whatever is in the sallyport and resumes CSSA 1, continuing
where we left off. CSSA 1 can execute additional code if it needs,
including additional syscalls. Once completed patching up for CSSA 0, it
exits successfully. The host then resumes CSSA 0.

This comes with some performance impact - usually for things that are IO
bound; so they are already slow. But the benefit is that we get to throw
away a significant amount of our hand-coded assembly. The result is
something far more maintainable.

Signed-off-by: Nathaniel McCallum <nathaniel@congru.us>