WIP: make this crate library that includes a binary
We want to allow other code (like a client/server) to use
`enarx_keepldr` as a library to build keeps, rather than having to
`exec()` the `enarx-keepldr` binary.

This commit refactors the crate to be a library - with an included
`enarx-keepldr` binary, built from `src/bin/enarx-keepldr.rs`.

This is a work-in-progress - the documentation on the `enarx_keepldr`
library is incomplete. There are some half-assed placeholder doc
comments here and there, but in order to get this to compile (so I can
make sure everything still works as expected) I've temporarily turned on
the #[allow(missing_docs)] directive.

Signed-off-by: Will Woods <will@profian.com>